+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - These files will still be included in a build of the simulator as that build process uses the Makefile
 - Disable BACnet routing as is currently not used in the device service
     - Because of this, we can now set our device ID as something other than 0
+- Updated Device Service object name
 
 ### Related to Simulator
 
@@ -16,3 +17,6 @@
 - Updated identifying values of device object
 - Updated handling logic of script and populate arguments
 - Updated default behaviour when no arguments provided where device instance is 1234 and it populates the sim with 1 of each object
+- Updated how the priority arrays are initialised for Analog Output, Binary Output and Binary Value
+- Removed any rules on setting certain priority levels
+- Updated device object name for simulator